Factor out derive code for `Resource`
